Upload form, validation
Сhanged the logic of validation of fields hashtags and description of the form.
Now validation occurs when characters are entered into the fields, and not when the form is submitted.

If the fields are not valid, button submit form is blocked and unblocked if the
fields are valid.

Created function onValidateFieldForm, which is called in function openPhotoEditingModal of module modal-upload.js.
It is responsible for validation of fields hashtags and description.