Social comments
- Create function hideCommentsLoaderButton (in comments.js module).
Depending on the number of comments, this function hides or shows the button for loading new comments.
Called in function createOnShowMoreComments

- Create function createOnShowMoreComments (in comments.js module).
This function creates click handlers for button commentsLoader (if the number of comments > 5);
It also removes click handler for button commentsLoader (if the number of displayed comments >= total number of comments);
If the number of comments is <5, it shows all comments and hides the button commentsLoader.

Called in function openFullPicture (full-picture.js module).